Move test, and switch to distributed.lock.Lock.
